Sereal::Decoder: fix compilation warning

+++ Perl/Decoder/srl_decoder.c
@@ -1083,7 +1083,7 @@ srl_read_hash(pTHX_ srl_decoder_t *dec, SV* into, U8 tag) {
         else
         if ( expect_false( SvTYPE(*fetched_sv) != SVt_NULL ) ) {
             /* sv_dump(*fetched_sv); */
-            SRL_RDR_ERRORf2(dec->pbuf, "duplicate key '%.*s' in hash", key_len, (char *)from);
+            SRL_RDR_ERRORf2(dec->pbuf, "duplicate key '%.*s' in hash", (int) key_len, (char *)from);
         }
         srl_read_single_value(aTHX_ dec, *fetched_sv, fetched_sv );
     }
